changeing doubles to longs for salary

+++ src/main/java/com/codeup/adlister/models/Ad.java
@@ -1,5 +1,5 @@
 package com.codeup.adlister.models;
-
+//ad class
 public class Ad {
     private long id;
     private long userId;
@@ -102,7 +102,7 @@ public class Ad {
         this.location = location;
     }
 
-    public double getSalary() {
+    public long getSalary() {
         return salary;
     }
